pyproject.toml
[project]
name = "risurealm-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "chromadb>=1.0.0",
    "fastapi>=0.115.0",
    "gradio>=5.0.0",
    "groq>=0.37.1",
    "httpx>=0.28.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
    "ty>=0.0.1a34",
]
